SRCNN/requirements.txt
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.0
Pillow>=8.0.0
matplotlib>=3.3.0
tqdm>=4.50.0
scikit-image>=0.17.2
tensorboard>=2.4.0
opencv-python>=4.5.0 